changed: int to float to rescale sound params, tests for audio functions, `rework read_audio` (#11)
* move: src files to src, create: dir for output files, create: dir for audio samples, fix: pathsin src files

* Applied changes requested from reviewers: pathlib, type annotations, type of returned value and moving samples and output files to .gitignore; also added dependencies

* changed: name of audio directory and name of microphone input file, added: returning np.array and sample rate from microphone input function

* Removing file saving inside microphone_input

* rename and lint

* Changed audio input from microphone to preserve proper scaling

* changed: int to float to rescale sound params

* Added tests for both file and microphone input parameters

* Added: file loading constraints(exceptions) and proper tests

* Changed: plotting is now controlled by flags

* Added: init.py in AI-dio Changed: read_sound() function for better code structure, plotting controlled by flags and rawdata output instead of log_mel

* Added: init.py in AI-dio Changed: read_sound() function for better code structure, plotting controlled by flags and rawdata output instead of log_mel

* Removed src. from imports

---------

+++ src/AI_dio/audio/microphone_input.py
@@ -16,7 +16,7 @@ def microphone_input(
 ) -> Tuple[np.ndarray, int]:
     logging.info("Recording...")
     audio: np.ndarray = sd.rec(
-        int(record_sec * rate), samplerate=rate, channels=channels, dtype="int16"
+        int(record_sec * rate), samplerate=rate, channels=channels, dtype="float32"
     )
     sd.wait()
     logging.info("Finished recording.")
